requirements.txt
celery==4.2.1
coreapi==2.3.3
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
factory-boy==2.12.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
redis==3.5.3
swagger-spec-validator==2.7.3
